+++ Changelog.md
@@ -41,6 +41,7 @@
   - [#377](../../issues/377) - ToggleButton highlight doesn't match the button size
   - [#388](../../issues/388) - QuickAccess item not visible when added from backstage
   - [#391](../../issues/391) - QAT items not displayed properly if tab is made visible after initialization
+  - [#392](../../issues/392) - Change Window Title doesn't update Title length
 
 - ### Enhancements
   - [#250](../../issues/250) - Enable change/hide of window title foreground